Signup and Login

+++ backend/controllers/sauces.js
@@ -1,25 +1,58 @@
 const Sauce = require('../models/Sauce');
+const fs = require('fs');
 
 exports.createSauce = (req, res, next) => {
-    delete req.body._id;
+    const sauceObject = JSON.parse(req.body.sauce);
+    delete sauceObject._id;
+    delete sauceObject._userId;
     const sauce = new Sauce({
-      ...req.body
+      ...sauceObject,
+      _userId: req.auth.userId,
+      imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
     });
+
     sauce.save()
-      .then(() => res.status(201).json({ message: 'Sauce enregistrée !'}))
+      .then(() => res.status(201).json({ message: 'Sauce enregistrée !' }))
       .catch(error => res.status(400).json({ error }));
 };
 
 exports.modifySauce = (req, res, next) => {
-    Sauce.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-      .then(() => res.status(200).json({ message: 'Sauce modifiée !'}))
+    const sauceObject = req.file ? {
+      ...JSON.parse(req.body.sauce),
+      imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    } : { ...req.body };
+
+    delete sauceObject._userId;
+    Sauce.findOne({ _id: req.params.id })
+      .then((sauce) => {
+        if (sauce.userId != req.auth.userId) {
+          res.status(401).json({ message: 'Non autorisé' })
+        }
+        else {
+          Sauce.updateOne({ _id: req.params.id}, { ...sauceObject, _id: req.params.id })
+            .then(() => res.status(200).json({ message: 'Sauce modifiée !' }))
+            .catch(error => res.status(401).json({ error }));
+        }
+      })
       .catch(error => res.status(400).json({ error }));
 };
 
 exports.deleteSauce =  (req, res, next) => {
-    Sauce.deleteOne({ _id: req.params.id })
-      .then(() => res.status(200).json({ message: 'Sauce supprimée !'}))
-      .catch(error => res.status(400).json({ error }));
+    Sauce.findOne({ _id: req.params.id })
+      .then(sauce => {
+        if (sauce.userId != req.auth.userId) {
+          res.status(401).json({ message: 'Non autorisé' });
+        }
+        else {
+          const filename = sauce.imageUrl.split('/images/')[1];
+          fs.unlink(`images/${filename}`, () => {
+            Sauce.deleteOne({ _id: req.params.id })
+              .then(() => res.status(200).json({ message: 'Sauce supprimée !' }))
+              .catch(error => res.status(401).json({ error }));
+          })
+        }
+      })
+      .catch(error => res.status(401).json({ error }));
 };
 
 exports.getOneSauce = (req, res, next) => {
